Modifies the way multimapping reads are filtered
Previously, the AS and XS scores would be compared.
Turns out that apparently bwa mem specifically sets the mapq score
to a value dependent on the number of equal hits.
To make a long story short, selecting for mapq > 3 is sufficient to get
rid of reads that have no equally good alternative alignment.