FileBased classes can specify a tmp file
If a temporary directory is specified the following will happen
depending on the mode:

- 'r' or 'r+': The file is copied to the temporary directory
- 'w', 'x', 'w-': The file is created in the temporary directory
- 'a': Either of the above depending on the existence of the file.

Two new methods have been added to FileBased to handle what happens when
the file is closed:

- finalize(): Copies the temporary file to the intended directory
- cleanup(): Deletes the temporary files

These two methods are automatically called on __exit__(), if the the
file was opened with a context manager (e.g. `with` statement), but have
to be explicitly called otherwise.